+++ test/unit/webrtc.test.ts
@@ -17,7 +17,7 @@ import { NamedAgent } from '../../src/types/named-agent';
 import { StanzaMediaSession } from '../../src/types/stanza-media-session';
 import { GenesysCloudMediaSession } from '../../src/types/genesys-cloud-media-session';
 import { IQ } from 'stanza/protocol';
-import { IMediaSession } from '../../src/types/media-session';
+import { IStanzaMediaSessionParams } from '../../src/types/media-session';
 
 jest.mock('../../src/types/stanza-media-session');
 StanzaMediaSession.prototype.on = jest.fn();
@@ -252,19 +252,17 @@ describe('prepareSession', () => {
     expect(StanzaMediaSession).toBeCalledWith(expect.objectContaining({ ignoreHostCandidatesFromRemote: true }));
   });
 
-  it('should delete pending session and sessionsMap tracking', () => {
+  it('should delete pending session', () => {
     (StanzaMediaSession as jest.Mock).mockReset();
     StanzaMediaSession.prototype.on = jest.fn();
     const client = new Client({});
     const webrtc = new WebrtcExtension(client as any, {} as any);
 
     webrtc.pendingSessions = { mysid: { sessionId: 'mysid' } as any };
-    webrtc['sessionsMap'] = { mysid: false };
 
     expect(Object.values(webrtc.pendingSessions).length).toBe(1);
     const session = webrtc.prepareSession({ sid: 'mysid' } as any);
     expect(Object.values(webrtc.pendingSessions).length).toBe(0);
-    expect(Object.values(webrtc['sessionsMap']).length).toBe(0);
   });
 
   it('should not delete pending session if sdpOverXmpp', () => {
@@ -328,6 +326,27 @@ describe('prepareSession', () => {
 
     const session = webrtc.prepareSession({} as any);
     expect(StanzaMediaSession).toBeCalledWith(expect.objectContaining({ ignoreHostCandidatesFromRemote: false }));
+  });
+
+  it('should update sessions and sessionsMap on terminated event', () => {
+    const client = new Client({});
+    const webrtc = new WebrtcExtension(client as any, {} as any);
+    (StanzaMediaSession as jest.Mock).mockReset();
+    (StanzaMediaSession as jest.Mock).mockImplementationOnce((params: IStanzaMediaSessionParams) => {
+      const emitter = new EventEmitter();
+      emitter['id'] = params.id;
+      return emitter;
+    });
+    webrtc['sessionsMap'] = { 'session25': false };
+
+    expect(webrtc.getAllSessions().length).toBe(0);
+    const session = webrtc.prepareSession({ sid: 'session25' } as any);
+    expect(webrtc.getAllSessions().length).toBe(1);
+
+    expect(Object.values(webrtc['sessionsMap']).length).toBe(1);
+    session?.emit('terminated');
+    expect(Object.values(webrtc['sessionsMap']).length).toBe(0);
+    expect(webrtc.getAllSessions().length).toBe(0);
   });
 });
 
@@ -724,6 +743,18 @@ describe('handledIncomingRtcSession', () => {
 });
 
 describe('initiateRtcSession', () => {
+  let client: Client;
+  let webrtc: WebrtcExtension;
+  let stanzaInstance: NamedAgent;
+
+  beforeEach(() => {
+    client = new Client({});
+    webrtc = new WebrtcExtension(client as any, {} as any);
+    stanzaInstance = getFakeStanzaClient();
+    stanzaInstance.jid = 'test@test.com';
+    webrtc['stanzaInstance'] = stanzaInstance;
+  });
+
   it('should add medias based on provided stream', async () => {
     const client = new Client({});
     const webrtc = new WebrtcExtension(client as any, {} as any);
@@ -825,6 +856,32 @@ describe('initiateRtcSession', () => {
 
     expect(sendSpy).toHaveBeenCalledWith('message', expected);
     expect(sendSpy).not.toHaveBeenCalledWith('presence', expect.any);
+  });
+
+  it('should send correct media presence for colaborateVideo', async () => {
+    const opts = {
+      jid: 'room@conference.test.com',
+      mediaPurpose: SessionTypes.collaborateVideo,
+      conversationId: 'test-conversation',
+      sourceCommunicationId: 'test-source'
+    };
+
+    const sendSpy = jest.spyOn(stanzaInstance, 'send');
+
+    await webrtc.initiateRtcSession(opts);
+
+    expect(sendSpy).toHaveBeenCalledWith('presence', expect.objectContaining({
+      type: 'upgradeMedia',
+      to: opts.jid,
+      from: webrtc.jid,
+      media: expect.objectContaining({
+        conversationId: opts.conversationId,
+        sourceCommunicationId: opts.sourceCommunicationId,
+        video: true,
+        audio: true,
+        videoGuest: true
+      })
+    }));
   });
 
   it('should handle when stream and params are provided', async () => {
@@ -1362,7 +1419,7 @@ describe('refreshIceServers', () => {
       await promise;
       fail('ice servers should have timedout');
     } catch (error) {
-      expect(error.message).toBe('Timeout waiting for refresh ice servers to finish');
+      expect((error as Error).message).toBe('Timeout waiting for refresh ice servers to finish');
     }
 
     jest.clearAllTimers();
@@ -1778,7 +1835,19 @@ describe('handleStanzaInstanceChange', () => {
     await webrtc.handleStanzaInstanceChange(stanza);
 
     expect(spy).toHaveBeenCalledWith(123);
-  })
+  });
+
+  it('should set existing sessions on the new Stanza instance', async () => {
+    webrtc['stanzaSessions'] = [
+      { id: 'session24' } as StanzaMediaSession,
+      { id: 'session25' } as StanzaMediaSession
+    ];
+
+    await webrtc.handleStanzaInstanceChange(stanza);
+
+    expect(Object.values(stanza.jingle.sessions).length).toBe(2);
+    expect(webrtc['stanzaSessions'][0].parent).toBe(stanza.jingle);
+  });
 });
 
 describe('getAllSessions', () => {
@@ -1796,7 +1865,7 @@ describe('getAllSessions', () => {
     const jSession1 = { id: '51' };
     const genSession1 = { id: '2' };
 
-    stanza.jingle.sessions = [ jSession1 ] as any;
+    webrtc['stanzaSessions'] = [ jSession1 ] as any;
     webrtc['webrtcSessions'] = [ genSession1 ] as any;
 
     const sessions = webrtc.getAllSessions();